[tech] Move 'impl_id' inside the 'transit_model_collection' crate
