+++ src/kv1/read.rs
@@ -28,7 +28,7 @@ use proj::Proj;
 use serde::Deserialize;
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::result::Result as StdResult;
-use transit_model_collection::{CollectionWithId, Id};
+use transit_model_collection::CollectionWithId;
 
 /// Deserialize kv1 string date (Y-m-d) to NaiveDate
 fn de_from_date_string<'de, D>(deserializer: D) -> StdResult<Date, D::Error>
